requirements.txt
streamlit==1.32.2
pandas==2.2.0
numpy==1.26.4
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.8
sqlalchemy==2.0.27
faiss-cpu==1.8.0
python-dotenv==1.0.1
openpyxl==3.1.2
scikit-learn==1.4.1.post1
plotly==5.19.0
fastapi==0.110.0
uvicorn==0.27.1
geopy==2.4.1
